fix: run boot recovery under the CLI entry point too
Boot recovery (trace-poll resume, orphan benchmark/evaluation-run
finalization) lived only in server/index.ts, which the CLI never
executes — 'agent-health serve' imports app.js and listens itself, so
recovery was dead code on the primary distribution path. Extract a
shared runBootRecoverySafely() and call it post-listen from both
entries (after AH_PORT reflects the bound port, since the trace poller
makes HTTP self-calls).
