+++ server/index.ts
@@ -15,10 +15,7 @@ import { createApp } from './app.js';
 import { getStorageConfigFromFile, getObservabilityConfigFromFile, getStorageConfigFromTs, getObservabilityConfigFromTs } from './services/configService.js';
 import { findObservioRoot, spawnObservioAgent, OBSERVIO_DEFAULT_PORT, resetObservioPort, isPortFree, setObservioPort, waitForObservioReady, killObservioAgent } from './services/observioAgent.js';
 import { validateAwsCredentials } from './services/tracesService.js';
-import { resumePendingTracePollsSafely } from './services/traceRecoveryOnBoot.js';
-import { recoverOrphanBenchmarkRunsSafely } from './services/benchmarkRunRecoveryOnBoot.js';
-import { recoverOrphanEvaluationRunsSafely } from './services/evaluationRunRecoveryOnBoot.js';
-import { getStorageModule } from './adapters/index.js';
+import { runBootRecoverySafely } from './services/bootRecovery.js';
 
 // Register server-side connectors (subprocess, claude-code)
 // This import has side effects that register connectors with the registry
@@ -159,25 +156,11 @@ async function startServer() {
         }
         console.log('');
 
-        // Resume orphan trace-mode polling that was lost during a restart.
-        // Fire-and-forget — must never block server startup or crash on failure.
-        // Runs after listen() so the poller's HTTP self-calls (asyncRunStorage)
-        // can reach the local API.
-        try {
-          const storage = getStorageModule();
-          if (storage) {
-            resumePendingTracePollsSafely(storage);
-            // Also fail out orphan BenchmarkRuns (status: 'running' for too long
-            // with the runner long dead). Different bug class — see
-            // server/services/benchmarkRunRecoveryOnBoot.ts.
-            recoverOrphanBenchmarkRunsSafely(storage);
-            // Top-level EvaluationRuns use a separate process-local registry;
-            // finalize any persisted `running` docs left by the prior process.
-            recoverOrphanEvaluationRunsSafely(storage);
-          }
-        } catch (err: any) {
-          console.warn(`[bootRecovery] Could not start: ${err?.message || err}`);
-        }
+        // Resume orphan trace polling / finalize orphan runs lost in a restart.
+        // Shared with the CLI entry point (cli/utils/startServer.ts) — see
+        // server/services/bootRecovery.ts. Runs after listen() so the trace
+        // poller's HTTP self-calls (asyncRunStorage) can reach the local API.
+        runBootRecoverySafely();
 
         // Graceful shutdown — stop background timers, kill child processes, drain connections
         const shutdown = (signal: string) => {
